trolley

+++ trolley/src/trolley.c
@@ -3,18 +3,18 @@
 #include<stdio.h>
 
 int main(){  //Συνάρτηση main
-long int l,r,left, right;  // Δήλωση μεταβλητών
+long long int l,r,left, right;  // Δήλωση μεταβλητών
 
 while(1){
 //Έλεγχος για το πρώτο κόστος
 		printf("Please enter the cost of going left:\n");
-	if((l=scanf("%ld",&left)) !=1 || l==EOF){
+	if((l=scanf("%lld",&left)) !=1 || l==EOF){
 		printf("Terminating.\n");
 return 0;	
 	}		
 //Έλεγχος για το δεύτερο κόστος
         	printf("Please enter the cost of going right:\n");
-	if((r=scanf("%ld",&right)) !=1 || r==EOF ){
+	if((r=scanf("%lld",&right)) !=1 || r==EOF ){
                 printf("No right cost provided.\n");
 return 1;
         }
